Refactor GeeklistService and add Async calls for Cards.
