+++ GeeklistSharp.Tests/GeeklistServiceHighfiveTest.cs
@@ -15,9 +15,6 @@ namespace GeeklistSharp.Tests
     [TestClass()]
     public class GeeklistServiceHighfiveTest : GeeklistBaseTest
     {
-        const string testCardId = "25c31dfce3d67208330a6cb995fc517bc48deda5d63bf6a65b83637cec65f9db";
-        const string testMicroId = "c8a1d5e5d41bbcb6d29f6b63b1e9e6526e78ee25a7a0983ec63ff8a4b2275148";
-
         private GeeklistService service;
 
         [TestInitialize]
@@ -30,11 +27,11 @@ namespace GeeklistSharp.Tests
         ///A test for ServiceHighfive to Highfive a given Card
         ///</summary>
         [TestMethod]
-        public void ServiceHighfiveCardTest()
+        public void HighfiveItemCardTest()
         {
             try
             {
-                var returnStatus = service.HighfiveItem(testCardId, Service.GeeklistItemType.Card);
+                var returnStatus = service.HighfiveItem(TestConstants.CARDID, Service.GeeklistItemType.Card);
                 Assert.IsNotNull(returnStatus);
             }
             catch (GeekListException gle)
@@ -47,11 +44,11 @@ namespace GeeklistSharp.Tests
         ///A test for ServiceHighfive to Highfive a given Micro
         ///</summary>
         [TestMethod]
-        public void ServiceHighfiveMicroTest()
+        public void HighfiveItemMicroTest()
         {
             try
             {
-                var returnStatus = service.HighfiveItem(testMicroId, Service.GeeklistItemType.Micro);
+                var returnStatus = service.HighfiveItem(TestConstants.MICROID, Service.GeeklistItemType.Micro);
                 Assert.IsNotNull(returnStatus);
             }
             catch (GeekListException gle)
