+++ GeeklistSharp.Tests/GeeklistServiceMicroTest.cs
@@ -27,7 +27,7 @@ namespace GeeklistSharp.Tests
         ///A test for ServiceMicros to get the Micros of the current User
         ///</summary>
         [TestMethod]
-        public void ServiceGetCurrentUsersMicrosTest()
+        public void GetCurrentUsersMicrosTest()
         {
             var currentUsersMicros = service.GetCurrentUsersMicros();
 
@@ -38,7 +38,7 @@ namespace GeeklistSharp.Tests
         ///A test for ServiceMicros to get the Micros of the current User
         ///</summary>
         [TestMethod]
-        public void ServiceGetCurrentUsersMicrosPagedTest()
+        public void GetCurrentUsersMicrosPagedTest()
         {
             var currentUsersMicros = service.GetCurrentUsersMicros(1, null);
 
@@ -49,7 +49,7 @@ namespace GeeklistSharp.Tests
         ///A test for ServiceMicros to get the Micros of the current User
         ///</summary>
         [TestMethod]
-        public void ServiceGetCurrentUsersMicrosCountTest()
+        public void GetCurrentUsersMicrosCountTest()
         {
 
             var currentUsersMicros = service.GetCurrentUsersMicros(null, 5);
@@ -61,7 +61,7 @@ namespace GeeklistSharp.Tests
         ///A test for ServiceMicros to get the Micros of the current User
         ///</summary>
         [TestMethod]
-        public void ServiceGetCurrentUsersMicrosPagedCountTest()
+        public void GetCurrentUsersMicrosPagedCountTest()
         {
 
             var currentUsersMicros = service.GetCurrentUsersMicros(2, 5);
@@ -73,7 +73,7 @@ namespace GeeklistSharp.Tests
         ///A test for ServiceMicros for a specific Micro
         ///</summary>
         [TestMethod]
-        public void ServiceGetSpecificMicroTest()
+        public void GetMicroTest()
         {
             var micro = service.GetMicro(TestConstants.MICROID);
 
@@ -84,7 +84,7 @@ namespace GeeklistSharp.Tests
         ///A test for ServiceMicros to create a new Micro
         ///</summary>
         [TestMethod]
-        public void ServiceCreateMicroTest()
+        public void CreateMicroTest()
         {
 
             var micro = service.CreateMicro("Unit Test Micro" + Guid.NewGuid()) as Micro;
@@ -96,7 +96,7 @@ namespace GeeklistSharp.Tests
         ///A test for ServiceMicros to create a new Micro
         ///</summary>
         [TestMethod]
-        public void ServiceCreateMicroForCardTest()
+        public void CreateMicroCardTest()
         {
 
             var micro = service.CreateMicro("card",TestConstants.CARDID,"Unit Test Micro" + Guid.NewGuid()) as Micro;
@@ -104,6 +104,7 @@ namespace GeeklistSharp.Tests
             Assert.IsNotNull(micro);
         }
 
+#region Micro Serialization Tests
         [TestMethod]
         public void DeserializeMicroTest()
         {
@@ -111,5 +112,6 @@ namespace GeeklistSharp.Tests
                 "{\"status\":\"ok\",\"data\":{\"status\":\"Unit Test Microe96f2127-7c2e-4936-8904-3ac566c02ae3\",\"slug\":\"2156\",\"permalink\":\"/lsmithmier/micro/2156\",\"trending_hist\":[],\"trending_at\":\"2012-02-10T20:23:03.833Z\",\"updated_at\":\"2012-02-10T20:23:03.872Z\",\"created_at\":\"2012-02-10T20:23:03.833Z\",\"reply\":{\"in_reply_to\":{\"id\":\"9b9ebefba8bc41329e5b865efd9f3124ff2d68f0a8c690ecbac2acd74f2841a7\",\"status\":\"Unit Test Microe96f2127-7c2e-4936-8904-3ac566c02ae3\",\"permalink\":\"/lsmithmier/micro/2156\",\"type\":\"micro\"},\"thread\":{\"id\":\"9b9ebefba8bc41329e5b865efd9f3124ff2d68f0a8c690ecbac2acd74f2841a7\",\"status\":\"Unit Test Microe96f2127-7c2e-4936-8904-3ac566c02ae3\",\"permalink\":\"/lsmithmier/micro/2156\",\"type\":\"micro\"}},\"is_active\":true,\"is_trending\":false,\"hashtags\":[],\"mentions\":[],\"stats\":{\"highfives\":0},\"short_code\":{\"id\":\"9b9ebefba8bc41329e5b865efd9f31247f090b9f9dd763b4662d1bbc5a918ab0\",\"gklst_url\":\"http://gkl.st/ubkXS\"},\"user\":{\"id\":\"e47163371a660b6eca2b7935ec31f058648175377b290fb039229b3a08971890\",\"screen_name\":\"lsmithmier\",\"avatar\":{\"small\":\"http://a2.twimg.com/profile_images/493173238/CSharp2_normal.png\",\"large\":\"http://a2.twimg.com/profile_images/493173238/CSharp2.png\"}},\"id\":\"9b9ebefba8bc41329e5b865efd9f3124ff2d68f0a8c690ecbac2acd74f2841a7\"}}";
 
         }
+#endregion Micro Serialization Tests
     }
 }
