+++ GeeklistSharp.sln
@@ -17,14 +17,19 @@ Global
 		CategoryFile = GeeklistSharp.vsmdi
 	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		ClientProfiles|Any CPU = ClientProfiles|Any CPU
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E4EF5C4E-708C-4934-B3C1-9D7CCF501D62}.ClientProfiles|Any CPU.ActiveCfg = Release|Any CPU
+		{E4EF5C4E-708C-4934-B3C1-9D7CCF501D62}.ClientProfiles|Any CPU.Build.0 = Release|Any CPU
 		{E4EF5C4E-708C-4934-B3C1-9D7CCF501D62}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E4EF5C4E-708C-4934-B3C1-9D7CCF501D62}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E4EF5C4E-708C-4934-B3C1-9D7CCF501D62}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E4EF5C4E-708C-4934-B3C1-9D7CCF501D62}.Release|Any CPU.Build.0 = Release|Any CPU
+		{364A93FE-4581-4A9D-849F-1A788A8F94AE}.ClientProfiles|Any CPU.ActiveCfg = Release|Any CPU
+		{364A93FE-4581-4A9D-849F-1A788A8F94AE}.ClientProfiles|Any CPU.Build.0 = Release|Any CPU
 		{364A93FE-4581-4A9D-849F-1A788A8F94AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{364A93FE-4581-4A9D-849F-1A788A8F94AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{364A93FE-4581-4A9D-849F-1A788A8F94AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
